backend/requirements.txt
django==4.2.29
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
pymysql==1.1.2
python-dotenv==1.0.1
